module udated

+++ init0.pp
@@ -1,3 +1,4 @@
+
 class docker {
 package { 'docker':
 ensure => present,
@@ -7,33 +8,42 @@ ensure => running,
 require => Package['docker'],
 }
 file { "/root/Dockerfile":
-        
+
         owner => 'root',
         group => 'root',
         source => 'puppet:///modules/docker/Dockerfile',
     }
 
 file { "/root/index.html":
-        
+
         owner => 'root',
         group => 'root',
         source => 'puppet:///modules/docker/index.html',
     }
+
 exec { 'rm':
 command => '/bin/docker rm -f $(sudo docker ps -a -q)',
-before => Exec['build'],
+#before => Exec['build'],
+#unless => '/bin/docker ps | grep website 1>/dev/null',
+#notify => Exec['build'],
+onlyif => '/bin/docker ps | grep website'
 }
+
 exec { 'build':
 command => '/bin/docker build . -t website',
+#require => Exec['rm'],
+#onlyif => '/bin/docker ps | grep website'
 }
 exec { 'run':
 command => '/bin/docker run -it -d -p 82:80 -d website',
 require => Exec['build'],
 }
+
+
 }
 
 class nrpe {
-	
+
 exec { 'epel-release':
 command => '/bin/yum install -y https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm',
 unless => '/bin/yum list installed | grep epel-release 2>/dev/null',
@@ -52,3 +62,4 @@ ensure => present,
 }
 
 }
+
